Added auto coloring and support for direct color
it is now possible to add a color attribute to labels.
all valid css formats are allowed (eg. "red" or "#FF0000").
the value is directly inserted into the css.
fixes #13

if there is a value without class and color attribute, it gets an automatic value.
this is done at compile time and the original file is overwritten.
fixes #14

also fixes minor bug in gulp file that created app/VersionInformation.tse .
this is now prevented.

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>