fix link colors and fixed z-indexes
z-indexes were wrong. It was possible to get the table header in behind of
labels. It was also possible that search options were behind the table header.
Labels were also in front of dialogs.
This is now resolved.
fixes #46

Some links had the wrong colors. This is now resolved, via restructuring the css-class
fixes #47

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>