Add style to unknown criteria values
Values of criteria that were not documented in table.json were almost not visible.
This path adds a black border and black text color to these labels.
The different style signifies creators that there are values that are not documented and the border
allows differentiating between different labels (eg. 'slow medium' and 'slow' + 'medium'
couldn't be differentiated)

solves #68

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>