Add support for deciding Development state on Repo
This commit adds the possibility to add columns that depend on a Repository.
These need to have a Repo attribute that is true in table.json.
You also have to add the criteria for the values (eg min-age 6 days and max-age 1 month).
The supported attributes are: min-age, max-age, min-age-unit, max-age-unit
  min-age (number):
    The minimum age to apply this label.
  min-age-unit (string):
    The unit of min-age. Supported units: d (days), m (months), y (years).
  max-age (number):
    The maximum age to apply this label.
  max-age-unit (string):
    The unit of max-age. Supported units: d (days), m (months), y (years).
Tooltips say how old the last commit is.
Up to now, only one unit is supported, so if you want 1.5 years, you have to use 18 months.

Future plans: add support for releases/tags and hours.

issue: #39
Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>