Exchange colors for pastel variants
Removed the first color (almost not visible as pastel) and transformed rest to pastel.
This was done via RGB -> HSL transformation and transforming HSL colors to pastel as
described in https://kentor.me/posts/generating-pastel-colors-for-css/
Tool for RGB->HSL: http://www.rapidtables.com/convert/color/rgb-to-hsv.htm

Because the white writing on pastel colors was almost not visible, I added a foreground
color (#808080) to elments that used auto coloring.

Issue: #79
Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>