Add number columns with range search
Added number columns/criteria that don't filter according to values but to ranges.
Filtering allows multiple ranges/values separated by commas.
An example would look like this: 100,200-250
Both min and max are inclusive.
Ranges max to min (eg 200-100) are handled as well.
NaN ranges are ignored.
It is also possible to add multiple numbers in a column and it tests if one is in range.

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>