Add expand and shrink option for table
The expand button adds all columns to the display, it is "<>".
The shrink button removes all columns that were added by the expand button, it is "><".
The function of the shrink button means that with a table consisting of the columns A, B, and C and A and B being the current display,
after the expand button was pressed, A, B, and C are shown.
If you remove B now you are left with A and C.
If you now shrink the table back, you are left with A because the shrink button disables all columns that were enabled through
the expand button.

fixes #90
Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>