Fix problem with empty bullet point
the empty bullet point was created by weird handling of spaces in filenames.
this problem is now mitigated by restructuring the docs directory.
at the same time, the template for ADRs was changed to MADRE.
the empty bullet point was created by weird handling of spaces in filenames.
this problem is now mitigated by restructuring the docs directory.
at the same time, the template for ADRs was changed to MADR.

solves #126

Fix fkeys.json generation for windows

Signed-off-by: Armin Hueneburg <hueneburg.armin@gmail.com>